refactor: group components and hooks within features
