+++ src/features/stake/hooks/useDetails.ts
