+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from 'next';
-import { Stake } from 'src/features/stake/Stake';
+import { Stake } from 'src/features/stake/components/Stake';
 
 const StakePage: NextPage = () => {
   return <Stake />;
